requirements.txt
astroid==2.3.3
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
emoji==0.5.4
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.19
pylint==2.4.4
python-dotenv==0.11.0
python-telegram-bot==12.4.2
pytz==2019.3
six==1.14.0
tornado==6.0.3
typed-ast==1.4.1
wrapt==1.11.2
